Checking in first version of a new indexer. This is written entirely in c, eliminating the dependency on ruby from the project this is forked from.
The first steps of the indexer are complete - eliminating the unneeded parts of the dump, and indexing the articles as they are archived. Then it walks through the archive and records the offsets for each article within its block. The next step is to find a small, searchable way to store the indicies.

This has been tested with the latest wiktionary dumps, and will soon be tested with wikipedia dumps. 

The dump and index this creates is not yet usable in the iPhone app - those changes are forthcoming, when the indexer is at a stable point.